feat/add cjs modules
